add attributes file to .src.mpm packages
the attributes file contains file attributes for the files that end up in
the binary package. format:

owner   group   mode    file

file can be any file that is in the installroot. filenames must start with
a leading /. "FILES" and "DIRECTORIES" (literal, and without a leading
slash) can be used to set defaults. rules are fired in order, so it is best
to list the defaults first.